pyproject.toml
[project]
name = "bayesian_network_dev"
requires-python = ">=3.12"
version = "0.0.0"
dependencies = [
#    "bayesian-network",
    "dynamic-bayesian-network",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "azure-artifacts"
url = "https://pypi.org/simple"
publish-url = "https://pkgs.dev.azure.com/mostertpim/BayesianNetwork/_packaging/BayesianNetwork/pypi/upload/"

[tool.uv.sources]
#bayesian-network = { workspace = true }
dynamic-bayesian-network = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "parameterized>=0.9.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "torchvision>=0.24.0",
]
